feature: type of the abrir

+++ 0x14-file_io/2-append_text_to_file.c
@@ -7,7 +7,7 @@
  */
 int append_text_to_file(const char *filename, char *text_content)
 {
-	ssize_t escribir;
+	int escribir;
 	int i, count, abrir;
 
 	if (filename == NULL)
